refactor(cli): extract shared scaffold logic into scaffold-shared
Move create's post-scaffold steps (push entities, deploy site, install
AI skills, print the project/dashboard/site summary) plus getTemplateById
and DEFAULT_TEMPLATE_ID into a new scaffold-shared.ts, so they can be
reused by sibling commands. Pure refactor — no behavior change.
