+++ packages/cli/src/cli/commands/project/create.ts
@@ -1,27 +1,24 @@
-import { basename, join, resolve } from "node:path";
+import { basename, resolve } from "node:path";
 import type { Option } from "@clack/prompts";
-import { confirm, group, isCancel, select, text } from "@clack/prompts";
+import { group, select, text } from "@clack/prompts";
 import { Argument, type Command } from "commander";
-import { execa } from "execa";
 import kebabCase from "lodash/kebabCase";
 import type { CLIContext, RunCommandResult } from "@/cli/types.js";
-import {
-  Base44Command,
-  getDashboardUrl,
-  onPromptCancel,
-  theme,
-} from "@/cli/utils/index.js";
+import { Base44Command, onPromptCancel, theme } from "@/cli/utils/index.js";
 import { InvalidInputError } from "@/core/errors.js";
-import { deploySite, isDirEmpty, pushEntities } from "@/core/index.js";
+import { isDirEmpty } from "@/core/index.js";
 import type { Template } from "@/core/project/index.js";
 import {
   createProjectFiles,
   listTemplates,
-  readProjectConfig,
   setAppConfig,
 } from "@/core/project/index.js";
-
-const DEFAULT_TEMPLATE_ID = "backend-only";
+import {
+  completeProjectSetup,
+  DEFAULT_TEMPLATE_ID,
+  getTemplateById,
+  printProjectSummary,
+} from "./scaffold-shared.js";
 
 interface CreateOptions {
   name?: string;
@@ -29,18 +26,6 @@ interface CreateOptions {
   template?: string;
   deploy?: boolean;
   skills?: boolean;
-}
-
-async function getTemplateById(templateId: string): Promise<Template> {
-  const templates = await listTemplates();
-  const template = templates.find((t) => t.id === templateId);
-  if (!template) {
-    const validIds = templates.map((t) => t.id).join(", ");
-    throw new InvalidInputError(`Template "${templateId}" not found.`, {
-      hints: [{ message: `Use one of: ${validIds}` }],
-    });
-  }
-  return template;
 }
 
 function validateNonInteractiveFlags(command: Command): void {
@@ -55,7 +40,7 @@ function validateNonInteractiveFlags(command: Command): void {
 
 async function createInteractive(
   options: CreateOptions,
-  ctx: Pick<CLIContext, "log" | "runTask">,
+  ctx: CLIContext,
 ): Promise<RunCommandResult> {
   const templates = await listTemplates();
   const templateOptions: Option<Template>[] = templates.map((t) => ({
@@ -116,7 +101,7 @@ async function createInteractive(
 
 async function createNonInteractive(
   options: CreateOptions,
-  ctx: Pick<CLIContext, "log" | "runTask">,
+  ctx: CLIContext,
 ): Promise<RunCommandResult> {
   ctx.log.info(`Creating a new project at ${resolve(options.path!)}`);
 
@@ -155,8 +140,9 @@ async function executeCreate(
     skills?: boolean;
     isInteractive: boolean;
   },
-  { log, runTask }: Pick<CLIContext, "log" | "runTask">,
+  ctx: CLIContext,
 ): Promise<RunCommandResult> {
+  const { log, runTask } = ctx;
   const name = rawName.trim();
   const resolvedPath = resolve(projectPath);
 
@@ -176,124 +162,19 @@ async function executeCreate(
     },
   );
 
-  // Set app config in cache for sync access to getDashboardUrl and getAppClient
   setAppConfig({ id: projectId, projectRoot: resolvedPath });
 
-  const { project, entities } = await readProjectConfig(resolvedPath);
-  let finalAppUrl: string | undefined;
-
-  if (entities.length > 0) {
-    let shouldPushEntities: boolean;
-
-    if (isInteractive) {
-      const result = await confirm({
-        message:
-          "Set up the backend data now? (This pushes the data models used by the template to Base44)",
-      });
-      shouldPushEntities = !isCancel(result) && result;
-    } else {
-      shouldPushEntities = !!deploy;
-    }
-
-    if (shouldPushEntities) {
-      await runTask(
-        `Pushing ${entities.length} data models to Base44...`,
-        async () => {
-          await pushEntities(entities);
-        },
-        {
-          successMessage: theme.colors.base44Orange(
-            "Data models pushed successfully",
-          ),
-          errorMessage: "Failed to push data models",
-        },
-      );
-    }
-  }
-
-  if (project.site) {
-    const { installCommand, buildCommand, outputDirectory } = project.site;
-
-    let shouldDeploy: boolean;
-
-    if (isInteractive) {
-      const result = await confirm({
-        message: "Would you like to deploy the site now? (Hosted on Base44)",
-      });
-      shouldDeploy = !isCancel(result) && result;
-    } else {
-      shouldDeploy = !!deploy;
-    }
-
-    if (shouldDeploy && installCommand && buildCommand && outputDirectory) {
-      const { appUrl } = await runTask(
-        "Installing dependencies...",
-        async (updateMessage) => {
-          await execa({ cwd: resolvedPath, shell: true })`${installCommand}`;
-
-          updateMessage("Building project...");
-          await execa({ cwd: resolvedPath, shell: true })`${buildCommand}`;
-
-          updateMessage("Deploying site...");
-          return await deploySite(join(resolvedPath, outputDirectory));
-        },
-        {
-          successMessage: theme.colors.base44Orange(
-            "Site deployed successfully",
-          ),
-          errorMessage: "Failed to deploy site",
-        },
-      );
-
-      finalAppUrl = appUrl;
-    }
-  }
-
-  // Add AI agent skills (--no-skills flag sets skills to false, otherwise defaults to true)
-  const shouldAddSkills = skills;
-
-  if (shouldAddSkills) {
-    try {
-      await runTask(
-        "Installing AI agent skills...",
-        async () => {
-          await execa("npx", ["-y", "skills", "add", "base44/skills", "-y"], {
-            cwd: resolvedPath,
-            shell: true,
-          });
-        },
-        {
-          successMessage: theme.colors.base44Orange(
-            "AI agent skills added successfully",
-          ),
-          errorMessage:
-            "Failed to add AI agent skills - you can add them later with: npx skills add base44/skills",
-        },
-      );
-    } catch {
-      // Skills installation is non-critical (e.g., user may not have git installed)
-      // The error message is already shown by runTask, so we just continue
-    }
-  }
-
-  log.message(
-    `${theme.styles.header("Project")}: ${theme.colors.base44Orange(name)}`,
+  const summary = await completeProjectSetup(
+    { projectId, name, resolvedPath, deploy, skills, isInteractive },
+    ctx,
   );
-  log.message(
-    `${theme.styles.header("Dashboard")}: ${theme.colors.links(getDashboardUrl(projectId))}`,
-  );
-
-  if (finalAppUrl) {
-    log.message(
-      `${theme.styles.header("Site")}: ${theme.colors.links(finalAppUrl)}`,
-    );
-  }
+  printProjectSummary(summary, log);
 
   return { outroMessage: "Your project is set up and ready to use" };
 }
 
 async function createAction(
-  { log, runTask, isNonInteractive }: CLIContext,
+  ctx: CLIContext,
   name: string | undefined,
   options: CreateOptions,
 ): Promise<RunCommandResult> {
@@ -303,7 +184,7 @@ async function createAction(
 
   const skipPrompts = !!(options.name ?? name) && !!options.path;
 
-  if (!skipPrompts && isNonInteractive) {
+  if (!skipPrompts && ctx.isNonInteractive) {
     throw new InvalidInputError(
       "Project name and --path are required in non-interactive mode",
       {
@@ -315,8 +196,6 @@ async function createAction(
       },
     );
   }
-
-  const ctx = { log, runTask };
 
   if (skipPrompts) {
     return await createNonInteractive(
